Implement data export functionality
Added JSON export capability with comprehensive error handling for backup/restore features. Refactored FileManager into a singleton pattern to be used within DataExporter, while being compatible with Android

- Added export confirmation dialog and error handling
- Converted write file operation to use Result pattern
- Updated FileManager to handle platform-specific file operations
- Added new localized strings for export operations

+++ composeApp/src/commonMain/kotlin/feature/settings/backup/data/ExpenseTaiyakiDataImporter.kt
@@ -84,7 +84,7 @@ private fun ExpenseDataSource.insertList(
             this.insert(
                 uuid = expense.uuid,
                 categoryId = category.uuid,
-                amount = expense.cost,
+                amount = expense.cost.toDouble(),
                 dateTime = localDateTimeToString(dateTime, A3DateFormat.ISO8601),
                 notes = expense.detail
             )
